Use size_t for corosync_service_engine.private_data_size. Unsigned int and size_t represent two different concepts

+++ include/corosync/coroapi.h
@@ -494,7 +494,7 @@ struct corosync_service_engine {
 	unsigned short priority; /* Lower priority are loaded first, unloaded last.
 				  * 0 is a special case which always loaded _and_ unloaded last
 				  */
-	unsigned int private_data_size;
+	size_t private_data_size;
 	enum cs_lib_flow_control flow_control;
 	enum cs_lib_allow_inquorate allow_inquorate;
 	char *(*exec_init_fn) (struct corosync_api_v1 *);
